refactor(slice5): duplication:scripts <0.5% — Restorchestrierung in clip-patch-cli + poolGrooveIds-Hoist
